CBSP-719 Defaults the number of concurrent downloads to 1 (#6)
* CBSP-719 Defaults the number of concurrent downloads to 1

The dowloads at a file level is already done in multiple
parts and as such can max out a connection. By downloading
multiple files at once, it slows down the download of the
indavila file. Which means in general the user has to way
longer before they can use the file. There is a setting to
control this behaviour

* Fixes style and moves strings to use format()